Moved getArgs and getUsersFromMention to utils
