+++ api/messages.js
@@ -12,7 +12,6 @@ const {
     Guild,
     ApplicationCommand,
     Client,
-    CommandInteraction,
     EmbedBuilder,
     ActionRowBuilder,
     ApplicationCommandBuilder,
@@ -26,7 +25,7 @@ const {
     APIActionRowComponent,
     APIApplicationCommand,
 } = require('discord.js');
-const { getSlashCommand} = require('./utils');
+const { getSlashCommand, getArgs } = require('./utils');
 const keys = require('../resources/languages/expanded/en_us.json');
 const { prefix } = require('../config.json');
 
@@ -771,57 +770,6 @@ function createActionRows(components) {
     return actionRows.length === 0 ? [currentRow] : actionRows;
 }
 
-/**
- * Gets a list of users from a string of mentions.
- * @param {Client} client - The client to use.
- * @param {string} mention - The string of mentions.
- * @returns {Promise<User[]>}
- */
-async function getUsersFromMention(client, mention) {
-    if(typeof mention !== 'string') return [];
-
-    const usersPattern = new RegExp(Discord.MessageMentions.UsersPattern.source, 'g');
-    const matches = mention.matchAll(usersPattern);
-    if(!matches) return [];
-
-    const userArray = [];
-    for(let match of matches) {
-        // match[0] = entire mention
-        // match[1] = Id
-        userArray.push(await client.users.fetch(match[1]));
-    }
-
-    return userArray;
-}
-
-/**
- * Gets an array of arguments from a CommandInteraction.
- * @param {Client} client - The client to use.
- * @param {CommandInteraction} interaction - The interaction to get the arguments from.
- * @returns {string[]}
- */
-function getArgs(client, interaction) {
-    if(!(interaction instanceof Discord.CommandInteraction)) return [];
-
-    const args = [];
-
-    function addArgs(option) {
-        if(option.type === Discord.ApplicationCommandOptionType.SubcommandGroup || option.type === Discord.ApplicationCommandOptionType.Subcommand) {
-            args.push(option.name);
-            option.options.forEach(opt => addArgs(opt));
-        }
-        else if(option.type === Discord.ApplicationCommandOptionType.Channel) args.push(option.channel);
-        else if(option.type === Discord.ApplicationCommandOptionType.User) args.push(option.user);
-        else if(option.type === Discord.ApplicationCommandOptionType.Role) args.push(option.role);
-        else if(option.type === Discord.ApplicationCommandOptionType.Attachment) args.push(option.attachment);
-        else args.push(option.value);
-    }
-
-    interaction.options.data.forEach(option => addArgs(option));
-
-    return args;
-}
-
 module.exports = {
     keys,
     ph,
@@ -835,6 +783,4 @@ module.exports = {
     getActionRows,
     getComponent,
     createActionRows,
-    getUsersFromMention,
-    getArgs,
 };
